# tsdb: faster postings sort with generic slices.Sort