--- a/tsdb/index/postings.go
+++ b/tsdb/index/postings.go
@@ -21,6 +21,7 @@ import (
 	"sync"
 
 	"github.com/pkg/errors"
+	"golang.org/x/exp/slices"
 
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/storage"
@@ -239,11 +240,9 @@ func (p *MemPostings) EnsureOrder() {
 
 	for i := 0; i < n; i++ {
 		go func() {
-			var sortable seriesRefSlice
 			for job := range workc {
 				for _, l := range *job {
-					sortable = l
-					sort.Sort(&sortable)
+					slices.Sort(l)
 				}
 
 				*job = (*job)[:0]
@@ -829,13 +828,6 @@ func (it *bigEndianPostings) Seek(x storage.SeriesRef) bool {
 func (it *bigEndianPostings) Err() error {
 	return nil
 }
-
-// seriesRefSlice attaches the methods of sort.Interface to []storage.SeriesRef, sorting in increasing order.
-type seriesRefSlice []storage.SeriesRef
-
-func (x seriesRefSlice) Len() int           { return len(x) }
-func (x seriesRefSlice) Less(i, j int) bool { return x[i] < x[j] }
-func (x seriesRefSlice) Swap(i, j int)      { x[i], x[j] = x[j], x[i] }
 
 // FindIntersectingPostings checks the intersection of p and candidates[i] for each i in candidates,
 // if intersection is non empty, then i is added to the indexes returned.