@bboreham 2022-09-08
CI is failing on Windows with an error I can't believe is connected to this change:
```
    testing.go:1097: TempDir RemoveAll cleanup: remove C:\Users\RUNNER~1\AppData\Local\Temp\TestReadCheckpointcompress=false1728930424\001\wal\00000030: The process cannot access the file because it is being used by another process.
--- FAIL: TestReadCheckpoint (1.97s)
    --- FAIL: TestReadCheckpoint/compress=false (1.82s)
```

This appears to the same as described in #10368.